requirements.txt
-i https://pypi.org/simple
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
cachetools==5.2.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
cssutils==2.4.2
databases==0.6.1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
emails==0.6
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jinja2==3.1.2
lxml==4.9.0
mako==1.2.3
markupsafe==2.1.1
passlib==1.7.4
premailer==3.10.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlalchemy==1.4.37
starlette==0.19.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
